% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Dynamic search should return the same results as the brute force algorithm when solving the knapsack problem.}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{a data frame which contains two variables v and w, stands for each items value and weight respectively.}

\item{W}{a postive integer which stands for the knapsack size}
}
\value{
the maximum knapsack value and which elements.
}
\description{
Dynamic search return the same results as brute force algorithm, but it should scale much better with O(W n) runing speed.
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <-
data.frame(
w=sample(1:4000, size = n, replace = TRUE),
v=runif(n = n, 0, 10000)
)
dynamic_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
\references{
Reference page link <https://en.wikipedia.org/wiki/Knapsack problem#0.2F1 knapsack problem>
}
