% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy search is to use the heurisitc or approximation for the knapsack problem.}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{a data frame which contains two variables v and w, stands for each items value and weight respectively.}

\item{W}{a postive integer which stands for the knapsack size}
}
\value{
the maximum knapsack value and which elements.
}
\description{
greedy search will not give an exact result, but it will reduce the computational complexity considerably.
}
\examples{
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:1200,], W = 2000)
}
\references{
Reference page link <https://en.wikipedia.org/wiki/Knapsack problem#Greedy approximation algorithm>
}
