% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack-package.R
\docType{package}
\name{knapsack-package}
\alias{knapsack}
\alias{knapsack-package}
\title{knapsack: This package is to solve knapsack problems with three different algorithms}
\description{
This package will contain three different functions for solving knapsack problem. The Knapasack problem is to find the knapsack with the largest value of the elements added to the knapsack.
}
\author{
\strong{Maintainer}: Xuan Developer \email{xuawa284@student.liu.se}

Other contributors:
\itemize{
  \item Ashen Developer \email{thawe276@student.liu.se} [contributor]
}

}
\keyword{internal}
